Store all images in status when ready
This gets updated with the current value of Spec images when the
overall service is in the Ready state.

Storing the current state of images in the Status will allow update
logic to be added which determines what update jobs to run, and what
images to propagate to the other controllers.